Beginning to transition elevstl to use new, modular code
