+++ src/LatLng.h
@@ -1,0 +1,14 @@
+#ifndef LATLNG_H
+#define LATLNG_H
+
+class Vector;
+
+class LatLng {
+  public:
+    float lat,lng;
+    LatLng (float,float);
+    void print();
+    Vector toCartesian(void);
+};
+
+#endif
